renamed useToast file to follow conventions
